feat: extract field processing module from main.ts
- Create src/field-processing/ directory structure
- Extract utility functions (getValueByPath, formatDateToYYYYMMDD, formatObjectDates)
- Extract default field constants for all tools
- Extract tool-specific field selection rules
- Extract main processing logic (_processSingleObjectInternal, selectFieldsFromData)
- Create central index.ts export point
- Update main.ts to use modular imports
- Maintain exact behavior and TypeScript compliance
- Enable parallel development without file conflicts

Resolves Task 2: Extract Field Processing Module
