+++ main.ts
@@ -17,8 +17,25 @@ import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
 import { z } from 'zod';
 import './database.js'; // Ensures database is initialized on startup
-import rateLimiter from './rate_limiter.js'; 
+import rateLimiter from './rate_limiter.js';
 import { zodToJsonSchema } from 'zod-to-json-schema';
+
+// Import field processing utilities
+import {
+  selectFieldsFromData,
+  GET_PROJECTS_DEFAULT_FIELDS,
+  GET_PROJECTS_TOOL_SPECIFIC_RULES,
+  GET_PROJECT_BY_ID_DEFAULT_FIELDS,
+  GET_PROJECT_BY_ID_TOOL_SPECIFIC_RULES,
+  GET_TASK_BY_ID_DEFAULT_FIELDS,
+  GET_TASK_BY_ID_TOOL_SPECIFIC_RULES,
+  GET_STATUSES_FIXED_FIELDS,
+  GET_USERS_ME_FIXED_FIELDS,
+  GET_USERS_FIXED_FIELDS,
+  GET_WORKSPACES_DEFAULT_FIELDS,
+  GET_TASKS_DEFAULT_FIELDS,
+  GET_TASKS_TOOL_SPECIFIC_RULES
+} from './src/field-processing/index.js';
 
 // Prune old records on server startup
 try {
@@ -36,362 +53,11 @@ const LOG_PRUNE_AFTER_MS = 600000; // 10 minutes in milliseconds
 
 let baseUrl = 'https://api.usemotion.com/v1'; // Base URL from Swagger spec or default
 
-// Helper function to retrieve a value from an object using a dot-notation path
-function getValueByPath(obj: any, path: string): { value: any; found: boolean } {
-  if (obj === null || obj === undefined || typeof obj !== 'object' || !path) {
-    return { value: undefined, found: false };
-  }
-  const parts = path.split('.');
-  let current = obj;
-  for (const part of parts) {
-    if (current === null || current === undefined || typeof current !== 'object' || !Object.prototype.hasOwnProperty.call(current, part)) {
-      return { value: undefined, found: false };
-    }
-    current = current[part];
-  }
-  return { value: current, found: true };
-}
 
-// Utility function to format a date string to YYYY-MM-DD
-// Returns original string if parsing fails or input is invalid, or null if input is null/undefined.
-function formatDateToYYYYMMDD(dateString: string | null | undefined): string | null {
-  if (dateString === null || dateString === undefined) {
-    return null;
-  }
-  try {
-    // Check if it's already in YYYY-MM-DD format to avoid re-processing
-    if (/^\d{4}-\d{2}-\d{2}$/.test(dateString)) {
-      return dateString;
-    }
-    const date = new Date(dateString);
-    // Check if date is valid after parsing
-    if (isNaN(date.getTime())) {
-      return dateString; // Return original if invalid date string
-    }
-    return date.toISOString().split('T')[0];
-  } catch (e) {
-    // In case of any other error during parsing or processing
-    return dateString; // Return original string on error as per PRD nuance
-  }
-}
 
-// Utility function to format specific date fields within an object to YYYY-MM-DD
-// Modifies the object in place if a field is found and validly formatted, 
-// otherwise leaves it as is or as formatted by formatDateToYYYYMMDD (which can return original on error).
-function formatObjectDates(obj: any, dateFieldPaths: string[]): void {
-  if (!obj || typeof obj !== 'object' || !dateFieldPaths || dateFieldPaths.length === 0) {
-    return; // No object or no fields to process
-  }
 
-  for (const path of dateFieldPaths) {
-    let current = obj;
-    const parts = path.split('.');
-    const lastPart = parts.pop(); // Get the actual field name and remove it from parts
 
-    if (!lastPart) continue; // Should not happen with valid paths
 
-    // Navigate to the parent of the target field
-    let parent = obj;
-    for (const part of parts) {
-      if (parent && typeof parent === 'object' && Object.prototype.hasOwnProperty.call(parent, part)) {
-        parent = parent[part];
-      } else {
-        parent = null; // Path is invalid or part not found
-        break;
-      }
-    }
-
-    // If parent is valid and has the target field, format it
-    if (parent && typeof parent === 'object' && Object.prototype.hasOwnProperty.call(parent, lastPart)) {
-      const originalValue = parent[lastPart];
-      const formattedDate = formatDateToYYYYMMDD(originalValue);
-      // formatDateToYYYYMMDD returns original on error/invalid, or null if input was null.
-      // We only update if it makes sense (e.g. if formatDateToYYYYMMDD actually changed it or it was null)
-      if (formattedDate !== originalValue || originalValue === null) { 
-        parent[lastPart] = formattedDate;
-      }
-    }
-  }
-}
-
-// Helper function to process a single object based on specified fields and rules
-function _processSingleObjectInternal(
-  obj: any,
-  fieldsToSelect: string[],
-  // toolSpecificRules will be used in later subtasks
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  toolSpecificRules?: Record<string, any>,
-  isDefaultMode?: boolean // Added for Subtask 1.3
-): any {
-  if (obj === null || obj === undefined || typeof obj !== 'object') {
-    return {}; // Or null, depending on desired behavior for non-objects
-  }
-  const result: Record<string, any> = {};
-  for (const fieldPath of fieldsToSelect) {
-    let currentValue: any;
-    let isValidField: boolean = false;
-
-    // MODIFICATION FOR SUBTASK 1.2:
-    if (fieldPath.includes('.')) {
-      const { value, found } = getValueByPath(obj, fieldPath);
-      if (found) {
-        currentValue = value; // Store for further processing
-        isValidField = true;
-      } else {
-        // Optionally set to null if not found, or omit. PRD implies omitting or standard error later.
-        // For now, let's omit if not found to keep responses clean.
-        isValidField = false;
-      }
-    } else if (Object.prototype.hasOwnProperty.call(obj, fieldPath)) {
-      currentValue = obj[fieldPath]; // Store for further processing
-      isValidField = true;
-    } else {
-      isValidField = false;
-    }
-
-    if (isValidField) {
-      if (Array.isArray(currentValue)) {
-        const itemSimplificationRule = toolSpecificRules?.[fieldPath]?.defaultItemSimplification;
-        
-        // Apply itemSimplificationRule if it exists, regardless of isDefaultMode, 
-        // when the array field itself (fieldPath) is being processed.
-        // This handles cases like PRD 2.3's "simplified optional" for assignees.
-        if (typeof itemSimplificationRule === 'function') {
-          try {
-            result[fieldPath] = currentValue.map(itemSimplificationRule);
-          } catch (e) {
-            console.error(`Error applying itemSimplificationRule for field '${fieldPath}':`, e);
-            result[fieldPath] = currentValue; // Fallback to full array on rule error
-          }
-        } else {
-          // No specific simplification rule for this array field (or it was handled by fields like array[].property).
-          // If !isDefaultMode, PRD 1.3.3 (return full array) applies unless a rule handled it.
-          // If isDefaultMode and no rule, also return full array (e.g. array of primitives).
-          result[fieldPath] = currentValue; 
-        }
-      } else if (typeof currentValue === 'object' && currentValue !== null && !isDefaultMode) {
-        // Optional Mode for an object field, check for container simplification (PRD 1.3.1)
-        const containerRule = toolSpecificRules?.[fieldPath]?.optionalContainerSimplification;
-        if (typeof containerRule === 'function') {
-          try {
-            const simplifiedObject = containerRule(currentValue);
-            Object.assign(result, simplifiedObject); // Merge properties from rule output
-          } catch (e) {
-            console.error(`Error applying optionalContainerSimplification for field '${fieldPath}':`, e);
-            // If rule fails, the original object might still be added if fieldPath was directly requested
-            // and no other rule applies. This part handles the rule itself failing.
-            // If fieldPath was a direct request (e.g. fields:["manager"]) and rule fails,
-            // it will fall through to the general 'else' block below if not already handled.
-            // To ensure it doesn't get added again by the generic logic if the rule was the SOLE source for these fields:
-            // We can check if result[fieldPath] was already set OR if the rule was supposed to be the exclusive handler for this fieldPath.
-            // For now, the Object.assign is simply skipped. If the fieldPath itself was requested, it might be added by the generic handler later.
-            // This behavior might need refinement based on how strictly PRD 1.3.1 should be interpreted on rule failure.
-          }
-        } else {
-          // Optional mode, object, but no specific container simplification rule.
-          result[fieldPath] = currentValue;
-        }
-      } else {
-        // Primitive value, or an object in default mode (no special container simplification for objects in default mode by default)
-        result[fieldPath] = currentValue;
-      }
-    }
-  }
-  // Apply post-processing rules after all fields are initially populated
-  for (const fieldKey in result) {
-    if (Object.prototype.hasOwnProperty.call(result, fieldKey)) {
-      const postProcessRule = toolSpecificRules?.[fieldKey]?.postProcessField;
-      if (typeof postProcessRule === 'function') {
-        try {
-          result[fieldKey] = postProcessRule(result[fieldKey]);
-        } catch (e) {
-          console.error(`Error applying postProcessField for field '${fieldKey}':`, e);
-          // On error, leave the original selected value in place
-        }
-      }
-    }
-  }
-  return result;
-}
-
-// Main utility function to select and transform fields from data
-function selectFieldsFromData(
-  data: any,
-  requestedFields?: string[],
-  toolDefaultFields?: string[],
-  toolSpecificRules?: Record<string, any>
-): any {
-  const effectiveFields = (requestedFields && requestedFields.length > 0) 
-    ? requestedFields 
-    : (toolDefaultFields && toolDefaultFields.length > 0 ? toolDefaultFields : []);
-  
-  const isDefaultMode = !(requestedFields && requestedFields.length > 0);
-
-  if (effectiveFields.length === 0 || data === null || data === undefined) {
-    return Array.isArray(data) ? [] : {};
-  }
-
-  if (Array.isArray(data)) {
-    return data.map(item => _processSingleObjectInternal(item, effectiveFields, toolSpecificRules, isDefaultMode));
-  } else {
-    return _processSingleObjectInternal(data, effectiveFields, toolSpecificRules, isDefaultMode);
-  }
-}
-
-const GET_PROJECTS_DEFAULT_FIELDS = [
-  'id',
-  'name',
-  'description',
-  'priorityLevel',
-  'dueDate', // Date formatting to be handled in Subtask 2.2
-  'status.name',
-  'completedTime', // Date formatting to be handled in Subtask 2.2
-  'taskCount'
-];
-
-const GET_PROJECTS_TOOL_SPECIFIC_RULES = {
-  'dueDate': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-  'completedTime': { postProcessField: (value: string | null | undefined) => value ? formatDateToYYYYMMDD(value) : null },
-  'manager': { // As per PRD 2.1 for get_projects
-    optionalContainerSimplification: (managerObj: any) => managerObj ? { 'manager.name': managerObj.name } : { 'manager.name': null }
-  }
-};
-
-// PRD 2.2: Recommended Default Fields for get_projects_by_projectId
-const GET_PROJECT_BY_ID_DEFAULT_FIELDS = [
-  'id',
-  'name',
-  'description',
-  'workspaceId',
-  'priorityLevel',
-  'dueDate',
-  'startDate',
-  'completedTime'
-];
-
-// PRD 2.2: Tool specific rules for get_projects_by_projectId
-const GET_PROJECT_BY_ID_TOOL_SPECIFIC_RULES = {
-  'dueDate': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-  'startDate': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-  'completedTime': { postProcessField: (value: string | null | undefined) => value ? formatDateToYYYYMMDD(value) : null },
-  // PRD 2.2: Optional Field Simplification Default for manager
-  'manager': { 
-    optionalContainerSimplification: (managerObj: any) => managerObj && managerObj.name ? { 'manager.name': managerObj.name } : { 'manager.name': null }
-  }
-};
-
-// PRD 2.3: Recommended Default Fields for get_tasks_by_taskId
-const GET_TASK_BY_ID_DEFAULT_FIELDS = [
-  'id',
-  'name',
-  'status.name', // Assuming status is an object with a name property
-  'priority',
-  'dueDate', 
-  'scheduledStart',
-  'scheduledEnd',
-  'duration',
-  'completed',
-  'project.name' // Assuming project is an object with a name property
-];
-
-// PRD 2.3: Tool specific rules for get_tasks_by_taskId
-const GET_TASK_BY_ID_TOOL_SPECIFIC_RULES = {
-  'dueDate': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-  // PRD 2.3: Optional Field Simplification Defaults
-  'creator': {
-    optionalContainerSimplification: (creatorObj: any) => creatorObj && creatorObj.name ? { 'creator.name': creatorObj.name } : { 'creator.name': null }
-  },
-  'workspace': {
-    optionalContainerSimplification: (workspaceObj: any) => workspaceObj && workspaceObj.name ? { 'workspace.name': workspaceObj.name } : { 'workspace.name': null }
-  },
-  'project': {
-    optionalContainerSimplification: (projectObj: any) => projectObj && projectObj.name ? { 'project.name': projectObj.name } : { 'project.name': null }
-  }
-  // Assignees array simplification will be added in Subtask 4.4
-};
-
-// PRD 2.4: Fixed fields for get_statuses
-const GET_STATUSES_FIXED_FIELDS = [
-  'name',
-  'isDefaultStatus',
-  'isResolvedStatus'
-];
-
-// PRD 2.5: Fixed fields for get_users_me
-const GET_USERS_ME_FIXED_FIELDS = [
-  'id',
-  'name',
-  'email'
-];
-
-// PRD 2.6: Fixed fields for get_users
-const GET_USERS_FIXED_FIELDS = [
-  'id',
-  'name',
-  'email'
-];
-
-// PRD 2.7: Default fields for get_workspaces
-const GET_WORKSPACES_DEFAULT_FIELDS = [
-  'id',
-  'name'
-];
-
-// PRD 2.3 / Task 13.1: Default fields for get_tasks, aligning with get_tasks_by_taskId
-const GET_TASKS_DEFAULT_FIELDS = [
-  'id',
-  'name',
-  'status.name', 
-  'priority',
-  'dueDate', 
-  'scheduledStart',
-  'scheduledEnd',
-  'duration',
-  'completed',
-  'project.name'
-];
-
-// Task 13.2: Tool specific rules for get_tasks
-const GET_TASKS_TOOL_SPECIFIC_RULES = {
-  // Date formatting rules for relevant fields present in a task object
-  'dueDate': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-  'completedTime': { postProcessField: (value: string | null | undefined) => value ? formatDateToYYYYMMDD(value) : null },
-  'startOn': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-  'scheduledStart': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-  'scheduledEnd': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-  'createdTime': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) }, 
-  'updatedTime': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) }, 
-  'lastInteractedTime': { postProcessField: (value: string | null | undefined) => formatDateToYYYYMMDD(value) },
-
-  // Optional container simplification rules (as per PRD 1.3.1 and 2.3)
-  'creator': {
-    optionalContainerSimplification: (creatorObj: any) => creatorObj && creatorObj.name ? { 'creator.name': creatorObj.name } : { 'creator.name': null }
-  },
-  'workspace': {
-    optionalContainerSimplification: (workspaceObj: any) => workspaceObj && workspaceObj.name ? { 'workspace.name': workspaceObj.name } : { 'workspace.name': null }
-  },
-  'project': {
-    optionalContainerSimplification: (projectObj: any) => projectObj && projectObj.name ? { 'project.name': projectObj.name } : { 'project.name': null }
-  },
-
-  // Default/Optional item simplification for arrays (PRD 1.2.2, 2.3)
-  'assignees': {
-    defaultItemSimplification: (assignee: any) => assignee && assignee.name ? { name: assignee.name } : {}
-  },
-  'chunks': {
-    defaultItemSimplification: (chunk: any) => {
-      if (!chunk) return {};
-      // Return all original properties of the chunk, but format dates within it
-      return {
-        ...chunk,
-        scheduledStart: formatDateToYYYYMMDD(chunk.scheduledStart),
-        scheduledEnd: formatDateToYYYYMMDD(chunk.scheduledEnd),
-        completedTime: chunk.completedTime ? formatDateToYYYYMMDD(chunk.completedTime) : null
-      };
-    }
-  }
-  // `labels` is an array of strings, so no item simplification needed.
-};
 
 const server = new McpServer({
   name: 'Motion AI Assistant',
